Rename DerivedPropertyBlueprint, add symbolic links
DerivedPropertyBlueprint is now named DerivedDescriptor.  All .meta
files found in core/meta are now pointers to their .mjson equivalents.
