+++ core/meta/blueprint-reference.meta
@@ -1,43 +1,1 @@
-{
-    "value_target_blueprint": {
-        "prototype": "core/meta/property-blueprint",
-        "properties": {
-            "name": "value",
-            "blueprint": {
-                "@": "root"
-            },
-            "mandatory": true,
-            "valueType": "object",
-            "valueObjectPrototypeName": "Blueprint",
-            "valueObjectModuleId": "core/meta/blueprint",
-            "helpKey": ""
-        }
-    },
-    "root": {
-        "prototype": "core/meta/module-blueprint",
-        "properties": {
-            "name": "BlueprintReference",
-            "customPrototype": false,
-            "propertyBlueprints": [
-                {
-                    "@": "value_target_blueprint"
-                }
-            ],
-            "propertyBlueprintGroups": {
-                "association-blueprint": [
-                    {
-                        "@": "value_target_blueprint"
-                    }
-                ]
-            },
-            "propertyValidationRules": {},
-            "blueprintModule": {
-                "%": "core/meta/blueprint-reference.meta"
-            },
-            "exportName": "BlueprintReference",
-            "module": {
-                "%": "core/meta/blueprint-reference"
-            }
-        }
-    }
-}
+object-descriptor-reference.mjson
