+++ core/meta/blueprint.meta
@@ -1,203 +1,1 @@
-{
-    "binder_reference": {
-        "prototype": "core/meta/blueprint-reference",
-        "properties": {
-            "valueReference": {
-                "blueprintName": "Binder",
-                "prototypeName": "Binder",
-                "blueprintModule": {
-                    "%": "core/meta/binder.meta"
-                }
-            }
-        }
-    },
-    "blueprint_properties_reference": {
-        "prototype": "core/meta/blueprint-reference",
-        "properties": {
-            "valueReference": {
-                "blueprintName": "PropertyBlueprint",
-                "prototypeName": "PropertyBlueprint",
-                "blueprintModule": {
-                    "%": "core/meta/property-blueprint.meta"
-                }
-            }
-        }
-    },
-    "property_validation_rules_reference": {
-        "prototype": "core/meta/blueprint-reference",
-        "properties": {
-            "valueReference": {
-                "blueprintName": "PropertyValidationRule",
-                "prototypeName": "PropertyValidationRule",
-                "blueprintModule": {
-                    "%": "core/meta/validation-rule.meta"
-                }
-            }
-        }
-    },
-    "blueprint_name": {
-        "prototype": "core/meta/property-blueprint",
-        "properties": {
-            "name": "name",
-            "blueprint": {
-                "@": "root"
-            },
-            "cardinality": 1,
-            "mandatory": true,
-            "denyDelete": false,
-            "readOnly": true,
-            "valueType": "string",
-            "enumValues": [],
-            "helpKey": ""
-        }
-    },
-    "blueprint_blueprintInstanceModuleId": {
-        "prototype": "core/meta/property-blueprint",
-        "properties": {
-            "name": "blueprintInstanceModuleId",
-            "blueprint": {
-                "@": "root"
-            },
-            "cardinality": 1,
-            "mandatory": true,
-            "denyDelete": false,
-            "readOnly": false,
-            "valueType": "string",
-            "enumValues": [],
-            "helpKey": ""
-        }
-    },
-    "blueprint_binder": {
-        "prototype": "core/meta/association-blueprint",
-        "properties": {
-            "name": "binder",
-            "blueprint": {
-                "@": "root"
-            },
-            "cardinality": 1,
-            "mandatory": false,
-            "denyDelete": false,
-            "readOnly": false,
-            "valueType": "object",
-            "enumValues": [],
-            "helpKey": "",
-            "targetBlueprint": {
-                "@": "binder_reference"
-            }
-        }
-    },
-    "blueprint_property_blueprints": {
-        "prototype": "core/meta/association-blueprint",
-        "properties": {
-            "name": "propertyBlueprints",
-            "blueprint": {
-                "@": "root"
-            },
-            "cardinality": -1,
-            "mandatory": false,
-            "denyDelete": false,
-            "readOnly": false,
-            "valueType": "set",
-            "enumValues": [],
-            "helpKey": "",
-            "targetBlueprint": {
-                "@": "blueprint_properties_reference"
-            }
-        }
-    },
-    "blueprint_property_blueprint_groups": {
-        "prototype": "core/meta/association-blueprint",
-        "properties": {
-            "name": "propertyBlueprintGroups",
-            "blueprint": {
-                "@": "root"
-            },
-            "cardinality": -1,
-            "mandatory": false,
-            "denyDelete": false,
-            "readOnly": false,
-            "valueType": "map",
-            "enumValues": [],
-            "helpKey": "",
-            "targetBlueprint": {
-                "@": "blueprint_properties_reference"
-            }
-        }
-    },
-    "blueprint_property_validation_rules": {
-        "prototype": "core/meta/association-blueprint",
-        "properties": {
-            "name": "propertyValidationRules",
-            "blueprint": {
-                "@": "root"
-            },
-            "cardinality": -1,
-            "mandatory": false,
-            "denyDelete": false,
-            "readOnly": false,
-            "valueType": "set",
-            "enumValues": [],
-            "helpKey": "",
-            "targetBlueprint": {
-                "@": "property_validation_rules_reference"
-            }
-        }
-    },
-    "root": {
-        "prototype": "core/meta/module-blueprint",
-        "properties": {
-            "name": "Blueprint",
-            "module": {
-                "%": "core/meta/blueprint"
-            },
-            "exportName": "Blueprint",
-            "customPrototype": false,
-            "propertyBlueprints": [
-                {
-                    "@": "blueprint_name"
-                },
-                {
-                    "@": "blueprint_blueprintInstanceModuleId"
-                },
-                {
-                    "@": "blueprint_binder"
-                },
-                {
-                    "@": "blueprint_property_blueprints"
-                },
-                {
-                    "@": "blueprint_property_blueprint_groups"
-                },
-                {
-                    "@": "blueprint_property_validation_rules"
-                }
-            ],
-            "propertyBlueprintGroups": {
-                "blueprint": [
-                    {
-                        "@": "blueprint_name"
-                    },
-                    {
-                        "@": "blueprint_blueprintInstanceModuleId"
-                    },
-                    {
-                        "@": "blueprint_binder"
-                    },
-                    {
-                        "@": "blueprint_property_blueprints"
-                    },
-                    {
-                        "@": "blueprint_property_blueprint_groups"
-                    },
-                    {
-                        "@": "blueprint_property_validation_rules"
-                    }
-                ]
-            },
-            "propertyValidationRules": {},
-            "blueprintModule": {
-                "%": "core/meta/blueprint.meta"
-            }
-        }
-    }
-}
+object-descriptor.mjson
