+++ core/meta/derived-property-blueprint.meta
@@ -1,104 +1,1 @@
-{
-    "derived_property_getterDefinition": {
-        "prototype": "core/meta/property-blueprint",
-        "properties": {
-            "name": "getterDefinition",
-            "blueprint": {
-                "@": "root"
-            },
-            "cardinality": 1,
-            "mandatory": false,
-            "denyDelete": false,
-            "readOnly": false,
-            "valueType": "string",
-            "enumValues": [],
-            "helpKey": ""
-        }
-    },
-    "derived_property_setterDefinition": {
-        "prototype": "core/meta/property-blueprint",
-        "properties": {
-            "name": "setterDefinition",
-            "blueprint": {
-                "@": "root"
-            },
-            "cardinality": 1,
-            "mandatory": false,
-            "denyDelete": false,
-            "readOnly": false,
-            "valueType": "string",
-            "enumValues": [],
-            "helpKey": ""
-        }
-    },
-    "derived_property_blueprint_dependencies": {
-        "prototype": "core/meta/property-blueprint",
-        "properties": {
-            "name": "valueType",
-            "blueprint": {
-                "@": "root"
-            },
-            "cardinality": -1,
-            "mandatory": false,
-            "denyDelete": false,
-            "readOnly": false,
-            "valueType": "string",
-            "enumValues": [],
-            "helpKey": ""
-        }
-    },
-    "property_blueprint_reference": {
-        "prototype": "core/meta/blueprint-reference",
-        "properties": {
-            "valueReference": {
-                "blueprintName": "PropertyBlueprint",
-                "prototypeName": "PropertyBlueprint",
-                "blueprintModule": {
-                    "%": "core/meta/property-blueprint.meta"
-                }
-            }
-        }
-    },
-    "root": {
-        "prototype": "core/meta/module-blueprint",
-        "properties": {
-            "name": "DerivedPropertyBlueprint",
-            "customPrototype": false,
-            "parent": {
-                "@": "property_blueprint_reference"
-            },
-            "propertyBlueprints": [
-                {
-                    "@": "derived_property_blueprint_dependencies"
-                },
-                {
-                    "@": "derived_property_getterDefinition"
-                },
-                {
-                    "@": "derived_property_setterDefinition"
-                }
-            ],
-            "propertyBlueprintGroups": {
-                "association-blueprint": [
-                    {
-                        "@": "derived_property_blueprint_dependencies"
-                    },
-                    {
-                        "@": "derived_property_getterDefinition"
-                    },
-                    {
-                        "@": "derived_property_setterDefinition"
-                    }
-                ]
-            },
-            "propertyValidationRules": {},
-            "blueprintModule": {
-                "%": "core/meta/derived-property-blueprint.meta"
-            },
-            "exportName": "DerivedPropertyBlueprint",
-            "module": {
-                "%": "core/meta/derived-property-blueprint"
-            }
-        }
-    }
-}
+derived-descriptor.mjson
