+++ core/meta/validation-rule.meta
@@ -1,128 +1,1 @@
-{
-    "blueprint_owner_reference": {
-        "prototype": "core/meta/blueprint-reference",
-        "properties": {
-            "valueReference": {
-                "blueprintName": "Blueprint",
-                "prototypeName": "Blueprint",
-                "blueprintModule": {
-                    "%": "core/meta/blueprint.meta"
-                }
-            }
-        }
-    },
-    "validation_rule_blueprint_name": {
-        "prototype": "core/meta/property-blueprint",
-        "properties": {
-            "name": "name",
-            "blueprint": {
-                "@": "root"
-            },
-            "cardinality": 1,
-            "mandatory": true,
-            "denyDelete": false,
-            "readOnly": true,
-            "valueType": "string",
-            "enumValues": [],
-            "helpKey": ""
-        }
-    },
-    "validation_rule_owner_blueprint": {
-        "prototype": "core/meta/association-blueprint",
-        "properties": {
-            "name": "blueprint",
-            "blueprint": {
-                "@": "root"
-            },
-            "cardinality": 1,
-            "mandatory": true,
-            "denyDelete": false,
-            "readOnly": true,
-            "valueType": "object",
-            "enumValues": [],
-            "helpKey": "",
-            "targetBlueprint": {
-                "@": "blueprint_owner_reference"
-            }
-        }
-    },
-    "validation_rule_blueprint_validationSelector": {
-        "prototype": "core/meta/property-blueprint",
-        "properties": {
-            "name": "validationSelector",
-            "blueprint": {
-                "@": "root"
-            },
-            "cardinality": 1,
-            "mandatory": false,
-            "denyDelete": false,
-            "readOnly": false,
-            "valueType": "object",
-            "valueObjectPrototypeName": "Selector",
-            "valueObjectModuleId": "core/selector",
-            "enumValues": [],
-            "helpKey": ""
-        }
-    },
-    "validation_rule_blueprint_messageKey": {
-        "prototype": "core/meta/property-blueprint",
-        "properties": {
-            "name": "messageKey",
-            "blueprint": {
-                "@": "root"
-            },
-            "cardinality": 1,
-            "mandatory": false,
-            "denyDelete": false,
-            "readOnly": false,
-            "valueType": "string",
-            "enumValues": [],
-            "helpKey": ""
-        }
-    },
-    "root": {
-        "prototype": "core/meta/module-blueprint",
-        "properties": {
-            "name": "PropertyValidationRule",
-            "customPrototype": false,
-            "propertyBlueprints": [
-                {
-                    "@": "validation_rule_blueprint_name"
-                },
-                {
-                    "@": "validation_rule_owner_blueprint"
-                },
-                {
-                    "@": "validation_rule_blueprint_validationSelector"
-                },
-                {
-                    "@": "validation_rule_blueprint_messageKey"
-                }
-            ],
-            "propertyBlueprintGroups": {
-                "validation-rule-blueprint": [
-                    {
-                        "@": "validation_rule_blueprint_name"
-                    },
-                    {
-                        "@": "validation_rule_owner_blueprint"
-                    },
-                    {
-                        "@": "validation_rule_blueprint_validationSelector"
-                    },
-                    {
-                        "@": "validation_rule_blueprint_messageKey"
-                    }
-                ]
-            },
-            "propertyValidationRules": {},
-            "blueprintModule": {
-                "%": "core/meta/validation-rule.meta"
-            },
-            "exportName": "ValidationRule",
-            "module": {
-                "%": "core/meta/validation-rule"
-            }
-        }
-    }
-}
+validation-rule.mjson
